+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sheet/web
 
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @schnsrw/casual-sheets@0.5.3
+
 ## 0.3.4
 
 ### Patch Changes